feat: Add estimated hours field to stories
This commit adds a new field called "estimated_hours" to the stories table in the database. The field is of type decimal and allows for up to 5 digits with 2 decimal places. It is nullable, meaning it can be left empty if no estimate is available. This field will be used to store the estimated time required to complete a story.