feat: Add redirection logic for customer stories
This commit adds logic to redirect the user to the appropriate URL when accessing customer stories. If the user is on the "resources/stories" page, they will be redirected to "resources/developer-stories". If the user is on a specific customer story page, their ID will be extracted from the URL and they will be redirected to "resources/developer-stories/{id}". This ensures a seamless transition for users accessing customer stories.