feat: Add TagAllProjects command and related models
This commit adds a new command called TagAllProjects, which creates a tag for each project and associates it with the project. It also introduces the Tag model, which is responsible for storing tag information. Additionally, a migration file is included to create the tags table and another migration file to create the taggables table for handling the many-to-many relationship between tags and entities.

The Tag model has a name attribute and uses polymorphic relationships to associate tags with different types of entities such as projects, apps, customers, and stories.

The Tag resource in Nova is also added, allowing administrators to manage tags through the Nova admin panel.

Overall, this commit enhances the tagging functionality in the application by providing a convenient way to tag all projects automatically.

chore: Remove unused fields and imports in Project.php

- Removed unused fields Tag, Number, and DateTime
- Removed unused imports for Panel and Tabs

feat: Add new story status "Waiting"

- Added a new story status called "Waiting" in the StoryStatus enum.
- Updated the Story model to include a morphToMany relationship with the Tag model.
- Added a belongsToMany relationship between the Story and Project models.
- Implemented a sendStatusUpdatedEmail method in the Story model to send email notifications when the status is updated.

feat: Add tags field to Story resource

This commit adds a new `tags` field to the `Story` resource in order to allow users to assign tags to stories. The `tags` field is added in the index, details, and edit views of the resource. Additionally, the commit modifies the `StoryShowedByCustomer` and `ArchivedStoryShowedByCustomer` classes to include the new `tags` field in their respective `$hideFields` arrays.

feat: Add tags functionality to App, Customer, and Project models

This commit adds the ability to associate tags with the App, Customer, and Project models. A new method `tags()` is added to each model to retrieve all the associated tags. Additionally, a `saved` event is registered for each model to automatically create a tag based on the model's name and save it. This allows for easier categorization and organization of these models based on their tags.

BREAKING CHANGE: None