feat: Add TaggableTypeFilter to Story resource in Nova
This commit adds a new filter called TaggableTypeFilter to the Story resource in Nova. The filter allows users to filter stories based on their associated tags. It includes methods for applying the filter to the query and retrieving the filter options from the database.

The changes include:
- Adding a new file, TaggableTypeFilter.php, which contains the implementation of the filter.
- Modifying the Story.php file to include an instance of TaggableTypeFilter in the filters array.

These changes enhance the filtering capabilities of the Story resource in Nova by allowing users to easily filter stories by tag type.