feat: Add support for filtering tag links by category
This commit adds functionality to filter tag links based on the category of the associated documentation. If the category is "Customer", only tags associated with customer documentation will be shown. If the category is "Internal", both internal and customer tags will be shown.

BREAKING CHANGE: This commit introduces a new model class, Documentation, which is used to retrieve the associated documentation for each tag.