feat: Enable sending reminder email to story creator
This commit enables the functionality to send a reminder email to the creator of a waiting story. The code previously had a commented out line for sending the email, which has now been uncommented and implemented using the `Mail` class. Additionally, the function `updteWaintingInStoryLog` is called after sending the email.