feat(nova): add CustomerTickets resource oc: 3796
This commit adds a new Nova resource called CustomerTickets. It includes the necessary code to define the resource, set its label and URI key, and configure authorization rules. Additionally, it defines filters for the resource and an index query to filter results based on user roles. The commit also includes the definition of three dynamic partition metrics for the resource: "Customer", "Status", and "Type". These metrics are used to display data in Nova cards.