feat: Add support for generating reports by customer and tag
This commit adds support for generating reports by customer and tag in the ReportController. The `generateReportByTag` and `generateReportByStatusTag` methods now accept an additional parameter `$customers`, which is used to filter the results based on the name of the user associated with each story. The table headers are also updated to include the names of the customers.