feat: add tab10-dev-type to reports
This commit adds a new tab, "tab10-dev-type", to the reports section. It includes the necessary changes in the ReportController.php file and adds a new blade template file for displaying the tab's content. The tab displays an analysis of developers by type, showing the total annual count as well as quarterly breakdowns.